applets/kicker: make listview widths more consistent
Kicker sizes its listviews (both menu-like and runner columns)
dynamically within a range, to allow for languages that tend to have
somewhat larger strings and to be more helpful in runner columns where
names tend to be longer (like files or bookmarks). The maximum length
for menu-like listviews is 1.5 times the default size, but for runner
columns it's double. This difference does not seem necessary, and having
some columns be twice as large as others is visually rather unbalanced.

We can just use the same 1.5 multiplier for both, simplify the code a
bit, and make the columns look a bit less unbalanced.