emojier: make .desktop file executable in kglobalaccel dir
Switch from FILES to PROGRAMS to make the installed file executable.

This is necessary to keep Emoji Selector working when built in a systemd
extension; in this case, it will be owned by you rather than root, and
needs to be executable or else KIO won't launch it from a global
shortcut.