applets/Kickoff: fix menu with long touch
Use a proper TapHAndler accepting only events from an anctual
touchscreen in order to show the context menu on long press.
On a mousearea is not possible to be reliable, because if the
actual mouse cursor position is outside the delegate, we will get
the synthetized mouse event reported with no button.

BUG:509898