kaccess: Fix screen reader sometimes not being able to be toggled off anymore
KAccessApp::readSettings() is called everytime a setting in the kcms/access
was changed and ok/apply of the kcms/access pressed.

Since we did re-setup the GlobalShortcut and connect the QAction::triggered
signal of the "Toggle Screen Reader On and Off" toggleScreenReaderAction
QAction in KAccessApp::readSettings() we then ended to receive >=2 times a
toggleScreenReaderAction's QAction::triggered signal when pressing 1 time the
META+ALT+S global shortcut after using kcms/access.

Result of that was that after using the kcms/access, it was not possible to
turn of the Orca Screen Reader anymore using the META+ALT+S global shortcut.

BUG: 505568
FIXED-IN: 6.5.0