kcms/dateandtime: Avoid showing warning for permission errors
Is a user enters the wrong password or cancels the prompt to enter a
polkit password, they have already been shown an error message.

We don't need to show a second prompt.

The method is changed to return after any timedate settings fail, as if
you can't unset NTP you can't set a time.

BUG: 501966