kcms/dateandtime: Drop non timedated compatible backend
The clock backend has two codepaths, one via the dbus interface of
timedate and a custom codepath via a manual KDE kauth process.

This manual process is fundamentally broken, when enabling NTP it runs
the ntp binary once. On the next rebot next time the KCM is open the
setting is checked but no NTP server is running.

Or potentially an NTP daemon is running but managed by a setup not known
to KDE, and the setting will be correctly unchecked.

Having a fallback at a KDE level doesn't work as we need integration
into system services. It's better to be clear and drop it.

Timedated interface providers include:
 - systemd
 - timedatex
 - openrc-settingsd

A mailing list thread on the topic can be found at:
https://mail.kde.org/pipermail/distributions/2023-July/001383.html