[activitymananger] Fix activity background image
The code looks into the appletsrc file to load the wallpaper for a given
activity.

However that only worked when a custom image was specified by full path.

When any of the built-in wallpapers is used the Image key instead points
to the wallpaper package folder, not any specific image.

Also when the default wallpaper is used the key is empty.

Fix all of these cases by selecting one image when given a folder, and
fall back to the default wallpaper (called Next) when no config is present.

Also when using a color background use the actual default color instead of black

BUG: 378693