Update TimezoneSelector selection when the timezone is reset through the KCM
Currently, the TimezoneSelector sets the timezone when the users selects
some place; this breaks the binding that was defined by the KCM, which means
that further updates from the KCM (such as resetting the value) are not
picked up. This commit makes sure that the selected time zone of the
TimezoneSelector is always in sync with the KCM value.

BUG:505916


(cherry picked from commit b04674affffe0f8a20424d60f1e9382c7de75f28)

Co-authored-by: Niccolò Venerandi <niccolo@venerandi.com>