Allow Dashboard to pre-select system settings result
The dashboard currently hardcodes that, if you search for an item, the very
first item of the very first grid will be selected. However, the first
grid might not be visible, if no entry from that grid applies to the search.
Instead, this commits selects the first item in the first visible grid.

BUG:506279


(cherry picked from commit 033dd69bddd820aef6a74850fddd5e9a14a2afcf)

Co-authored-by: Niccolò Venerandi <niccolo@venerandi.com>