Make cell height in widget explorer as tall as the tallest element within it
Previously, the cell height of the widget explorer was hardcoded. This resulted
in text overflowing on certain fonts (such as Noto Sans Arabic 10pt).
Instead, we make sure that the cellHeight is the max of the height of the
AppletDelegates within it.

BUG:434814


(cherry picked from commit cf4e21f1033565533085a50caa2a3cfaf0d491db)

Co-authored-by: Niccolò Venerandi <niccolo@venerandi.com>