FolderModel: set applet as parent
If there is no parent, and we open the KNewFileDialog inside a popup
view, it crashes due to a dangling pointer.

Making sure we always have a parent if there is an applet
set to the model fixes this by ensuring there is an object
for the KNewFileDialog to point to.