FolderView: Disable mouse interaction in edit mode
In edit mode user is dragging widgets and other items. We want
to show where the items are so that user doesn't drag widget on
top of them, but we do not want the user to accidentally
move or open any desktop items.

This patch sets the mouse listener to enabled based on editmode