Focus searchbox when opening Kickoff
Some input methods such as compose key require the search box
to be focused in order to work. Previously, kickoff would open
with the focus to the gridview, and the searchbox would only get
focused when the user started typing. After this commit, the
focus is immediately given to the searchbox so that the
compose key can work.
Note that this means that the top row has to be navigated
through tab/alt+tab instead of left+right arrows, since the
latter has to be reserved to move the selection in the grid,
even if the searchbox is focused.

BUG:459269