applets/kicker: fix arrow down from search field to column
Amends 8746d41e8c4480621955b0d95b3a22e922eddb09

When displaying runner columns, Kicker initially highlights the first
entry of the first column even while focus is still on the search field,
as pressing return will activate that entry immediately. Previously,
pressing down from the search field would skip the already highlighted
entry and move to the second element in the first column. During the
keyboard navigation refactoring, this special case was not correctly
handled. This makes pressing arrow down feel as if nothing happens, as
the selection does not change (something does happen, focus moves from
the search field to the search results, but this is a distinction with
little immediate difference).

This change adds special handling for this case, restoring behavior back
to how it worked for 6.4.