sddm-theme: Avoid rendering invisible contents
An opacity of 0 will still render the items, particularly the
blinking password field cursor will still be active.

This was addressed for the lock screen in 45e0a722 of
plasma-workspace but not for the sddm-theme.

While the mentioned commit also set "enabled" to avoid
some alleged cursor blinking, I could not reproduce this
and just "visible" false seems to work fine.

Signed-off-by: Kiriakos Antoniadis <kiriakos.antoniadis@advantest.com>