applets/kicker: allow context menus with only favoriteId
Amends d5648c19293b617d6c6fa72ce724265bb1eae2cb

Some menu entries in kicker have no actionList, but do have a favoriteId
and can be added or removed from the list of favorites. In particular,
this applies to system actions. This case was not properly handled in
the above commit, so it was not possible to add system actions to the
favorites sidebar.

Instead, we allow showing the context menu even if the actionList is
empty, as long as a favoriteId is present.

BUG: 511448
FIXED-IN: 6.5.3