[kcms/keyboard] Don't reconfigure on unrelated config changes
KConfigWatcher can trigger on unrelated changed, e.g. when something
notifies on kdeglobals.

This resulted in configureKeyboard() being called, which has the
sideffect of switching to the first configured layout, disregarding the
current one.

To avoid this check which group has changed and only process the
settings when the right one has changed