applets/kicker: explicitly set sidebar layout width
Some styles will break when the layout width is not explicitly set, and
push the sidebar buttons outside the frame (or even the applet itself,
cutting them off). This includes a small number of third-party styles
(possibly only Diamond), but also Oxygen. The issue seems to be somewhat
inconsistent, and in some cases only applies when starting Plasma, not
when switching to an affected theme from breeze.

We can just explicitly set it and center the sections, which resolves
the issue. The alignment is not quite optimal with Oxygen, but should at
least be less obviously bad.

BUG: 511411
CCBUG: 505847