applets/kicker: use better sizing for sidebar
When the sidebar items were changed into toolbuttons to facilitate
keyboard accessibility, the sizing was intentionally cut rather close to
keep the spacings otherwise identical. This causes issues with some
Plasma styles, including Oxygen and Diamond, and causes the focus/hover
highlight to feel rather cramped in Breeze as it is very close to the
sidebar frame.

Instead of pre-calculating a size, we let the component's own sizing
bubble up so that the button width is always style-appropriate, use
the intended sidebar frame margins from the style as well, and add some
more explicit layouting instruction. This marginally increases the
sidebar size, but looks less cramped when items are highlighted, and
should resolves issues with some themes.

BUG: 511411
BUG: 505847