applets/kickoff: block hovering until mouse moves
Kickoff already has an implementation of hover blocking for searching:
once the user starts typing, it will not automatically select and focus
the item under the mouse, so keyboard navigation behaves consistently
no matter where the mouse is (unless the user accidentally moves it).
The same is not true for when the applet initially opens; then the
selected item changes depending on the mouse position, and users cannot
build muscle memory to launch favorites - e.g. opening Kickoff with a
keyboard shortcut and immediately pressing Return will not always launch
the same favorite.

We can just move the existing hover block, which is specific to the
search results, directly into the FullRepresentation, and adjust things
slightly to set the correct internal state depending on the context.
This makes hover blocking work for all of (expanded) Kickoff.