applets/kickoff: don't move focus from search bar when hovering
When an item becomes the currentItem through hovering, it takes
activeFocus. This can interfere with keyboard usage in the case of
pointer drift or accidental movements, as pressing Space will activate
the item now while the user may have intended to add a space character
to the search string. Moving focus away from the search field isn't
strictly necessary in this case - activation with Return and even moving
the selected item with the mouse and then continuing to move the
selection with arrow keys continues to work without. Always forwarding
Space presses to the search field instead, as kicker does currently,
breaks consistency with other applets and apps that use space to
activate focused controls.

Instead, we don't move activeFocus to the hovered item if the search
field currently has it. This allows the best of both worlds - users can
still use arrow keys followed by Space to activate (as in most applets
and apps) but unintended mouse movements will no longer launch
applications as the user tries to add a space character to the search
term.