applets/taskmanager: make highlightWindows label more precise
The highlightWindows option in the task manager config dialog currently
reads "... when hovering over previews". But this is not necessarily
correct; the user can disable previews in the tooltips and the option
will still work, only now the user can hover over the whole tooltip
(section) to trigger it.

So we make the text depend on the other setting, and if previews
are disabled, make it read "... when hovering over tooltips" instead.