applets/taskmanager: do less unnecessary work when previews are disabled
We don't need to be loading thumbnails and PipeWire stuff for previews
that won't be shown anyway.

Noticed while scrubbing over Task Manager tasks while the qpwgraph app
happened to be open (really useful tool!).

The tooltips are still fairly heavy, but this helps a bit for the
"thumbnails are disabled" use case.