desktoppackage/configuration: Adjust About page label positions
The About page in the widget configuration dialog has two lines in a
ColumnLayout displaying the widget name and description; the height of
this is determined by the icon next to them. Both lines share the
available space equally due to Layout.fillHeight, but the actual
position of the labels ends up looking rather odd, as one is at the very
top and the other at the very bottom, so there's a large gap between
them, and they don't seem well-related to the icon.

Instead, we place the top row at the bottom of its available space,
and the bottom row at the top. This makes the two lines look centered
around the icon, and gives adequate spacing between the lines.