applets/kickoff: "fix" footer overflow menu
In Plasma 6.5 with Qt 6.10, if the action button row is too long to fit
the widget, it no longer hides them in the overflow menu, but has them
overlap the tab bar or even go outside the widget. The cause seems truly
bizarre and has to be a qtbug; a simple assignment with an or expression
causes one of the variables used as input to become undefined... but
only if it is on the right-hand side of ||, LHS is fine. So
after "const C = A | B", B is somehow now undefined, while with
"const C = B | A" everything is fine. The issue seems to happen on at
least Fedora, KDE  Linux, and Arch, according to personal experience and
Bugzilla reports.

This clearly doesn't make any sense and needs to be fixed in qt
(qtbug-141920), but until then we can "fix" the symptoms by just
swapping the LHS and RHS.

BUG: 511503


(cherry picked from commit b74225ad48412d15b2cfb10d32ef2778c455aead)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>