desktoppackage/configuration: adjust About page text styling
The About page header in widget configuration dialogs currently has only
the heading at level 2 to distinguish it from the widget name. The
Kirigami Addons version of the dialog instead puts it at level 3 and
also sets the type to secondary heading, which gives it less visual
prominence and a clearer separation from the widget name, which serves
as the title. Ideally, these dialogs should be completely merged, but
in the mean time we can tweak it a little to bring their appearance more
in sync.

This change makes the heading text styling match Kirigami Addons'
(FormCard) About page.