applets/kicker: fix premature "No matches" placeholder
The visibility of the no matches placeholder depends on the current
search string not being empty and either not having a search running
or being already visible, to reduce visual flickering if the user
adds characters to a search string with no results. This logic doesn't
work consistently, as both the search running test and the search string
test are triggered by the search string's changed signal, and so their
order is arbitrary. If the visibility condition is evaluated before the
searchRunning property is updated, it will display and keep the
placeholder even while a search is running.

Instead, the placeholder now keeps a copy of the search string and tests
against that. This way, we can make sure we get the timing right and
update the searchRunning property first.


(cherry picked from commit 98fa028a3fc93897e64cddc29217ddb276f89e0e)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>