Folder: preserve icon layout on sorted mode after drag and drop
When we're in sorted mode (e.g. sorted by name) and we perform drag
and drop operation, our unsorted icon layout is overriden by a sorted
layout.

Previously after drag and drop operation we get back to old unsorted
layout, so it was impossible to make use of sorted layout to reorganize
our icons. Moreover it was a bit confusing that moving an icon changed
entire layout.

We can still go back to our unsorted layout as long as we don't perform
drag and drop operation on desktop.