applets/keyboardlayout: don't (attempt to) hide shortcut config
The keyboard layout's General config page contains code to hide the
default keyboard shortcuts page. This was introduced in 2002 (098a965b)
to work around the keyboard shortcut not working properly on the applet
and the intended method for setting a layout switching shortcut is in
kcm_keyboard. However, this is not a good idea:

- to do so, it requires mucking around in internals of how the config
  dialog framework is organized. Changes there may break it, and it's
  not easy to keep this in sync. In fact, it doesn't work on currently
  released Plasma (for who knows how long) and no one noticed.
- tooling doesn't like it because it requires looking up things by name
  during runtime
- it only applies to the widget configuration dialog even if it worked,
  so you could still set a keyboard shortcut for a tray widget through
  the system tray Entries config page or through kcm_keys anyway
- assigning and using a keyboard shortcut actually works as expected
  now, the original bug that commit intended to fix no longer applies.

The code is no longer necessary and does nothing except cause tooling
to complain, so let's kill it.