applets/kicker: remove justOpenedTimer
Kicker's justOpenedTimer runs for 600ms after expanding (or opening the
dash) and has three effects that apply while it's running:
- disable the active highlight on the panel (Kirigami.)icon (only active
  if the icon is hovered). This was added in 2015 to work around the
  highlight no longer being active after opening and closing the dash,
  but this would work without now (and even better as the highlight
  appears immediately, rather than after a 600 ms delay)
- disable the custom trianglemousefilter on the kicker menu
- disable the custom trianglemousefiler on the dash's sidebar

None of these seem particularly worth it, eepecially given the short
time that they are active and how the trianglemousefilter mostly doesn't
do anything even if you explictly try to get it to, but it does add
complexity due to adding timer state, with unclear and weak effects,
spread in various parts of kicker. So let's get rid of it.