applets/taskmanager: drop dead code for middle button close
In 2019, special code was added to not resize the task manager buttons
after middle mouse button closing them while the mouse is still hovering
the task manager (2861ae7c). Most of this code was removed as part of
the port to using a GridLayout, but some parts remain. This includes (a)
a signal handler that would set a property when something was closed
with the middle mouse button while hovering, but it doesn't work because
it's full of tests that no longer work, (b) that property which is never
used for anything, and (c) a property that stores the winId of tasks
closed with middle mouse button, which is not used for anything except
the test in (a), and I don't think it can ever get far enough to
actually use it because the test breaks before it gets there.

Even if we wanted to add this feature back (if necessary and feasible),
the little that remains is so broken that starting there doesn't seem
like a good idea, and while it's here, it's confusing and wastes
resources. So let's get rid of the remnants.