applets/kicker: fix subsubmenu opening direction on far edge panel
Kicker flips its orientation partially when on a right-edge panel in ltr
or left-edge panel in rtl. The submenus should open in the matching
direction, but do not - the submenu is forced to be correct by the
screen edge, but any subsubmenus are not and will open as usual for
the LayoutDirection, i. e. covering the main Kicker menu, rather than
away from it as for top/bottom/near edge panels.

Support for fixing this was added to p-w, so we enable it here and solve
the issue. Along the way we fix another bug where the text in actual RtL
languages may be left-aligned rather than right-aligned.