ScreenMapper: fix icon placement on multiple containments
Move icons to disabled screens when screen doesn't exist. The screen
is most likely not yet initialized when Plasma is loading, so do the
same logic as when removing screen to preserve icon layout.

Change screen mapping when swapping screens.

Remove consistency map, because it disallows to place all icons on
screen with number different than 0.

BUG: 512381