applets/kicker: simplify column/delegate widths
Kicker does lots of manual size calulations for delegates, listviews,
and their widths. This is somewhat cumbersome and makes the workings
harder to understand and change.

Instead, we can switch this over to using layouts and have Qt do the
work for us. We still percolate sizes upward to the listview, bounded by
the maximum and minimum widths, but properly reset them now, so that if
the column width will not persist in an increased state longer than
necessary.

BUG: 507728
FIXED-IN: 6.5.0