applets/window-list: add translation contexts and fix string
This commit adds translation context to the few strings still missing
them, and makes a string translatable that was not before. We also
avoid feeding a translated string as an argument into another translatable
string, as constructing it fresh allows the word order to be different in
the full sentence if necessary.