applets/kicker: use automatic sizing with layouts
Kicker uses a good bit of complicated manual sizing, which is
error-prone and difficult to keep in sync with modifications.

We can instead convert a few things to Layouts, add a few implicit
sizes where necessary, let the sizes bubble up instead of trying to
determine them ahead of time, and have Qt figure out almost all the
sizing for us. This simplifies the code a lot and makes it more easily
modifiable.