applets/kicker: drop MouseEventListener in ItemListView
ItemListView uses a MouseEventListener mostly to track hover focus
state. The same thing works with a HoverHandler, so we can just use that
and get rid of one level of indentation and use a more standard
component at the same time. While we're at it, we also get rid of the
hover handler for Runner columns - they don't actually do anything,
and it's not clear in the first place what the code is actually trying
to do and whether that's a good idea - the hover handler for the items
themselves already takes care of everything (I suspect it's a leftover
from a previous version that set focus instead of activeFocus). We also
remove the mouseMoved property, which is no longer necessary with the
new hover-blocking implementation.