applets/kicker: get rid of RunnerListViewContainer
The ItemListViews in the runner columns are wrapped in a container item,
to support the setting to allow the results to align to the bottom or
the top. We don't actually need this and it makes sizing a bit more
complicated as it's another layer to keep in mind.

So we remove it. We do a bit of trickery to support the option by
setting Layout.alignment, but further simplification of ItemListView
should also allow us to get rid of this. As the hoverable part of the
column is now larger, we make the delegate set activeFocus on the list,
rather than the list itself.