applets/kicker: remove ItemListView FocusScope
ItemListView has a FocusScope as its base. Its child is a ScrollView
(also a FocusScope) containing a ListView (also a FocusScope). That's
more focus being scoped than necessary. We can reduce one level of
indentation, and also use ScrollView's built-in capabilities as a
control to remove a HoverHandler and some properties. With this
simplified structure, we can also move even more of the sizing to
using Qt's automatic layouts, and in particular get rid of most manual
clamping in favor of Layout.minimum/maximum properties.