applets/kicker: remove RunnerResultsList FocusScope
RunnerResultsList is wrapped in a FocusScope. The only contained item
taking focus is a (Item)ListView, which itself is a FocusScope. We can
remove one level of indentation and some forwarding of sizes by getting
rid ofit. The only hurdle is that we now need to be slightly more
explicit about passing focus to the ItemListView, rather than just the
whole column, but that is all centralized in one place so it's
straightforward to implement.