applets/kicker: switch to individual Keys handlers
Kicker (and kickerdash) currently have most of their keyboard event
handling in often rather large Keys.onPressed handlers. This is
difficult to read, and requires manually accepting the event,which is
easy to forget and hard to spot.

Instead, we split it up into handlers for individual keys, extracting
things that apply to multiple keys (such as left/right arrow handling,
which may do either job depend on the layoutDirection) into smaller
functions that only contain the parts relevant to those keys. This
only requires manually un-accepting the event in case a key event needs
to be explicitly passed to the parent, which is rare, and the resulting
code has better separation and is thus more readable.