applets/kicker: simplify window handling
Kicker does a song and dance whenever a submenu is opened, turning off
hideOnWindowDeactivate before and then turning it on again in various
ways. This does not seem to be necessary; a Plasma Dialog opening does
not cause hideOnWindowDeactivate to trigger (not even on X11). Seems
like we can just remove it. This also means that we don't need
windowSystem's focusIn triggers anymore. The windowSystem's hidden
triggers were already only used to cause a reset in the MenuRep when
it is closed, this does not seem to work anymore, but it doesn't need to
as there's also an onExpandedChanged trigger that does the same thing
anyway (and actually works!).

Removing all this *almost* allows us to get rid of windowSystem;
unfortunately X11 still needs forceActive as otherwise the submenus
do not get focus with keyboard navigation. The relevant lines
now contain Plasma 6.8 TODOs; once we no longer need to worry about that
we can kill them here (and remove the whole class from p-w's kicker)