applets/taskmanager: Fix last-used task selection
The LastActivated property is undefined until a task is activated at least once. So when
no task in the group was ever activated, it can't find the top one using LastActivated.
This regression was introduced by !3307, and can be reproduced by clicking on a grouped
task right after restarting a running plasmashell.

To fix this, default to StackingOrder in that case. 

BUG: 513019