applets/kicker: use regular tooltips
Kickerdash's ItemGridDelegate uses a TooltipArea with a custom
mainItem to display its tooltips. While this works, it's different
in both positioning and sizing from the regular tooltips for similar
elements in e.g. kickoff; and the structure is more complex to
understand as it has more moving parts in different areas of the
kicker code.

Instead, we switch this to a simple PlasmaComponents ToolTip and a
HoverHandler (and even that one can be dropped should this ever be
ported to an ItemDelegate or other Control, rather than a base Item).

And while we're at it, we port the SideBarItem tooltips for regular
kicker from attached tooltips to regular PC3 tooltips, as the attached
version does not use the correct Plasma styling.