applets/kicker: move sizing SvgItems to DashboardRepresentation
Regular kicker no longer pre-computes the sizes based on the theme
metrics, but lets the actual sizes bubble up. The items used to
determine the theme margins are now only required for kickerdash. So we
can move them there and skip instantiating them for regular kicker, and
also skip a couple of linter warnings as it can now look it up by id
from the same file in most circumstances.