desktoppackage/activitymanager: cleanup some dead code
ActivityManager contains a function that is never called from anywhere,
and which itself does nothing except call another function that doesn't
exist. There's also a loader for a new activity dialog that never has a
source set; creating new activities happens through the kcm now. The
loader is not used for anything except adjusting a button's opacity if
it's ready, which it never will be as there's nothing to load in the
first place.

This change removes the function and the loader.