emojier/app: merge CategoryAction into Emoijier.qml
The delegate for creating category actions is its own small qml file.
Some parts of the action require access to data from the caller, which
means unqualified lookups or passing lots of extra information to the
delegate. Some of what is there is also essentially duplicated (such
as setting a string property category that is exactly the same as the
already existing property modelData.

It's easier, clearer, and less code to just merge the delegate into the
single place where it's used.