emojier: make screenreader-accessible
Emojier's category page sets default focus on the search field and sets
a forwardTo on the grid view to pass arrow keys and other interaction
keys that should operate on the view or delegate there. This works, but
means that focus always remains on the search field, and screen readers
are not informed about the currently selected item, giving no feedback
to the user and letting the explicitly set accessible properties on the
delegate go to waste.

Instead, we switch to regular handling where the delegate actually
receives focus, and set a forwardTo in the other direction, so that keys
the view/delgate do not handle are passed to the search field, allowing
modification of the search string from outside the search field. We
replace the remnant of a key passing mechanism, as using a forward
allows us to centralize the search modification behavior in one place,
rather than having two implementations that work differently. We also
make the Clear History button (sometimes) next to the search field
arrow key-accessible.