emojier/app: properly separate copy handling
Emojier's handling of copying is implemented in three different places
spread across both source files. As we want to modify the recents model
and access the window to show passive notifications, it makes sense to
handle this all centrally in the main app, and have the active page
signal it as needed. This allows for a less direct coupling of intent
and implementation, and helps reduce the coupling between both
components while reducing implementation duplication.