emojier/app: centralize setting data set and page title
The model, category, and page title of the CategoryPages are all
set from Emojier.qml... except when the user searches in a restricted
emoji set and the app switches to full search, then it happens from
inside CategoryPage. This means that the full model will have to be
either be accessed through an unqualified lookup from there or
explicitly passed to the page, and feels like improper separation of
concerns.

Instead, we provide a signal by which the page can request being
switched to the full data set, and handle all cases of setting the
data set and page title in one place, Emojier.qml.