emojier/app: use less confusing collapsible binding.
The globalDrawer's collapsible property is bound to
!topContent.activeFocus. topContent is not an item but a list of
QObjects, so it does not have an activeFocus property, so this
is just !undefined. It's also not clear what this is even supposed
to do, the app does not set any topContent for the globalDrawer so
there's nothing in the list that could have activeFocus.

Just set it to always true.