emojier/app: don't create unnecessary new pages.
All page loading actions in Emojier are implemented as replacing the
current page with a new one. This causes a warning on startup (as we're
trying to replace a page on an empty pageStack) and is generally not
necessary, as all the pages are the same, they just differ in some
properties.

So we can just create the page once and have the switching actions swap
out the relevant bits in that page, as we already do when a page is
switching to searching all data when nothing is found in the selected
category. This resolves the runtime warnings, saves the overhead of
having to create new pages, and allows a small interface convenience -
the user can now type a search term and then click the sidebar to choose
a category and only show results from that category. Before, the user's
search term would be cleared by the new page being loaded.