emojier/app: use regular highlight for currentItem
The emoji GridView currently uses the highlighted delegate style for the
currentItem; in QQC2 a solid color. We typically use this style to show
some persistent state (e.g. the currently open page from a sidebar with
delegates), and for  simple selection of items in a list or grid we use
the regular view highlight (in QQC2, currently a solid outline with a
transparent background).

This change switches emojier to the second style.