containments/desktop: add active property
Drag&Drop handling for FolderView checks the active property of the back
button to see if it's supposed to forward the drag&drop finished event
there. Unfortunately, there's no such property so this doesn't do
anything. I'm not 100% certain what was intended to happen here exactly,
but I think it's supposed to alias the running property of the internal
timer, and that does something useful - if the user tries to drop
something on the back button, it will no longer trigger the hover
activation of the button after the user already dropped the file on it.
(Dropping the file doesn't work with or without this change, but at
least attempting to do so doesn't also change the directory, which is
progress I guess...)