applets/kicker: don't open empty context menu
Right-clicking an ItemListDelegate will always attempt to show a context
menu, even if the context menu is empty. This will cause any submenus
visually parented to the delegate to close (and likely reopen a bit
later as the pointer is still over the delegate). This is pointless
visual noise, so let's just skip opening the context menu if there's
nothing to show.