applets/kicker: use abstraction for closing kicker(dash)
Various actions in kicker conclude the interaction with the widget
(such as launching an app, but not adding it to the favorites). To
implement this, the applets modify the PlasmoidItem directly, But
this way has several disadvantages: first, the PlasmoidItem needs to be
accessed from everywhere; this means either passing it everywhere or
unqualified lookups. Second, the specific way that that the widget is
closed is an implementation detail that the components do not need to
know anything about, and in fact works differently in kicker and
kickerdash (as one is a regular expanded widget and the other a custom
window). This makes it harder to share components between both widgets,
and is confusing - in fact, in one place kickerdash tries to hide itself
as if it were regular kicker.

Instead we abstract away from this using a signal that can be emitted
whereever relevant, and is passed up to the PlasmoidItem, where the
actual implementation lives. This makes the code less tightly coupled
and reduces the number of unqualified lookups.