emojier: add skin tone filter
The 'People and Body' category, is with 2261 emojis, by far the category
with the most emojis. Most of these are variants in skin tone.
By hiding 1615 of these variants and by showing only one type of variant
at a time, it's easier to discover emojis in the category/all view.