containments/panel: fix applet context menu order
The popup menu for panel applets is supposed to reorder itself so that
"Remove" is not the topmost action on topmost panels, as it is a
destructive action that should be placed away from the initial cursor
position if possible. To do so, the config overlay accesses the location
property of the popup. This is always undefined, however, as the
property doesn't actually exist. Instead, we need to access the
popupDirection property, with a flipped edge as this property is the
direction of the popup relative to the parent.

We implement this, and while we're at it, use a LayoutItemProxy so we
don't have the almost exact same code twice.