views/Desktop: Set applets property explicitly to list of strings
This ensures that the applets property sends a notify of its change
properly whenever it's changed.

With "var" type, it will not trigger updates to any bindings
using it. With "list<string>" type the update is sent properly.
See: https://doc.qt.io/qt-6/qml-var.html#change-notification-semantics

Also use "let" instead of "var" in AppletDelegate onClicked, to ensure
the variables are in correct scope and properly reset.

BUG: 511729