applets/kicker: replace custom Dashboard filter list delegate
DashboardRepresentation's sidebar has a custom delegate that is
essentially just an ItemAbstractDelegate with a custom contentItem.
It also still uses the same not-very-functional triangle mouse filter
that kicker used for opening submenus.

Instead, we replace it with ItemListDelegate, and use the same simple
timer to delay switching the category a little, like regular kicker does
now. This cuts a lot of unnecessary duplication and increases the
interaction consistency between different launchers.