applets/kicker: make dashboard highlight span delegate
The highlight for the dashboard sidebar list items is restricted to the
label width. This is misleading, as the non-highlighted area will still
activate on hover and allow right-clicking (for the few entries that
have a context menu). It's also not how our list views usually behave,
where the highlight spans the whole delegate.

Instead, remove all the complicated positioning and sizing, and just
use the highlight stock - it does the right thing on its own.