applets/kicker: unify delegates
Kicker contains three delegates that largely share the same structure
and code: the list delegates and sidebar delegates in main kicker, and
the grid delegates in kickerdash (the filter list sidebar in kickerdash
would be a fourth one, but that can essentially be replaced with the
list delegate from main kicker). Currently, the code is fully duplicated
across them.

Instead, we create a new base delegate class, and derive the delegate
components from that. This allows a substantial reduction in code
duplication, and makes further changes simpler.