applets/kicker: don't reset kickerdash model after drag
Kickerdash resets the root model whenever a drag operation is finished,
whether successfully or canceled. This is to work around bug involving
some unspecified pointer grab issues; the code is ten years old, marked
FIXME TODO HACK and is unsure of the cause of the thing, promising a
future intense debug session that hasn't happened yet. The problem with
doing this is that the main grid's model is set imperatively, and
resetting resetting the root model will leave the child models as null,
and the new model being created will not be used automatically as there
is no binding that would do it for us. (The All Apps view still works,
but appears to produce crashing for some people in this situation that
I can't reproduce, and will spew lots of runtime errors due to some
things being null). This is similar to what was fixed for kickoff in
f8532d35a782af807448fde581f3f7239c77b7d6, only this time it triggers on
any drag operation, only this time it triggers on any drag&drop
operation rather than when installing software in the background.

To fix this, we could employ the same strategy as in that commit and
rebind the model on refresh (and probably at some point should, to
handle the case of installing apps). But the question is whether the
original hack is still necessary ten years later - removing it appears
to produce no pointer issues and everything works as it should; it seems
likely that other changes in Qt or kicker(dash) in the last 10 years
made it unnecessary. So let's remove it.

CCBUG: 510655