applets/taskmanager: use weaker type for PlasmaPa GlobalConfig
QML is having trouble mapping the correct property type here, and fails
loading the optional PulseAudio integration.

PlasmaPa does not use declarative typie registration yet, so the
stronger type doesn't seem to be all that helpful in the first place
(and it's only accessed in one place anyway).

Let's use the weaker QObject type for now to restore functionality.