applets/taskmanager: Port audio handling to C++
The current code for the audio handler piggy-backs
on plasma-pa and is done entirely in QML.

Not only does this create an awkward dependency on plasma-pa,
the code is also spread out and hard to follow.

Move most of it to C++, leveraging PulseAudioQt. This centralizes
the implementation in a new AudioInfo class that is instantiated from
QML and exposes all relevant information