containments/panel: avoid unqualified lookups for root
ConfigOverlay accesses the root object from main.qml for various things
using unqualified lookups.

Instead, we identify the things that it needs from there, and pass them
in explicitly. For dragAndDropping, which is barely used in main and set
and read often in ConfigOverlay, we invert this and take the value from
ConfigOverlay where necessary.